Close #280, Refactor and resolve failed tests in api wallet and matching gateway (#282)
* wallet: Use data classes in models

* wallet: Use samples data

* wallet: [WIP] Refactor wallet manager

* wallet: Use samples in tests

* wallet: Refactor tests

* wallet: Resolve all failed tests

* matching-gateway: [WIP] Add samples object

* Use mockK in tests

* Convert all doubles to decimal

* matching-gateway: Resolve failed tests

* wallet: Refactor tests

* api: Resolve failed tests

* wallet: Remove user limits

* accountant: Fix failed tests

* GitHub: Add testing step to workflows

+++ accountant/accountant-core/src/main/kotlin/co/nilin/opex/accountant/core/model/PairFeeConfig.kt
@@ -1,9 +1,11 @@
 package co.nilin.opex.accountant.core.model
 
+import java.math.BigDecimal
+
 class PairFeeConfig(
     val pairConfig: PairConfig,
     val direction: String,
     val userLevel: String,
-    val makerFee: Double,
-    val takerFee: Double
+    val makerFee: BigDecimal,
+    val takerFee: BigDecimal
 )
