+++ bc-gateway/bc-gateway-ports/bc-gateway-eventlistener-kafka/src/main/kotlin/co/nilin/opex/bcgateway/ports/kafka/listener/model/EditCurrencyEvent.kt
@@ -1,7 +1,9 @@
 package co.nilin.opex.bcgateway.ports.kafka.listener.model
 
+import java.math.BigDecimal
+
 data class EditCurrencyEvent(
     val name: String,
     val symbol: String,
-    val precision: Double
+    val precision: BigDecimal
 ) : AdminEvent()
