+++ websocket/websocket-core/src/main/kotlin/co/nilin/opex/websocket/core/inout/CandleData.kt
@@ -1,17 +1,18 @@
 package co.nilin.opex.websocket.core.inout
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class CandleData(
     val openTime: LocalDateTime,
     val closeTime: LocalDateTime,
-    val open: Double,
-    val close: Double,
-    val high: Double,
-    val low: Double,
-    val volume: Double,
-    val quoteAssetVolume: Double,
+    val open: BigDecimal,
+    val close: BigDecimal,
+    val high: BigDecimal,
+    val low: BigDecimal,
+    val volume: BigDecimal,
+    val quoteAssetVolume: BigDecimal,
     val trades: Int,
-    val takerBuyBaseAssetVolume: Double,
-    val takerBuyQuoteAssetVolume: Double,
+    val takerBuyBaseAssetVolume: BigDecimal,
+    val takerBuyQuoteAssetVolume: BigDecimal,
 )
